# src/main/kotlin/com/example/payment_service/dto/res/SettlementBatchRes.kt
package com.example.payment_service.dto.res

import java.math.BigDecimal

data class SettlementBatchResponse(
    val batchRef: String,
    val merchantId: String,
    val totalAmount: BigDecimal,
    val feeDeducted: BigDecimal,
    val transactionCount: Int
)